added hamburger and appbar component
