+++ src/App.js
@@ -8,16 +8,10 @@ import Form from "./pages/form";
 import NotFound from "./pages/not-found";
 
 import { makeStyles } from "@material-ui/core/styles";
-import HomeIcon from '@material-ui/icons/Home';
-import PeopleIcon from '@material-ui/icons/People';
-import CreditCardIcon from '@material-ui/icons/CreditCard';
-import SettingsIcon from '@material-ui/icons/Settings';
 
 import './App.css';
 
 import CssBaseline from "@material-ui/core/CssBaseline";
-
-import {Sidebar, SidebarItem} from './components/sidebar';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -30,14 +24,6 @@ function App() {
   return (
     <div className={classes.root}>
       <CssBaseline />
-      <Sidebar>
-        <SidebarItem label="Home" icon={HomeIcon} link="/" />
-        <SidebarItem label="Form Example" icon={PeopleIcon} link="/form"/>
-        <SidebarItem label="Settings" icon={SettingsIcon}>
-          <SidebarItem label="Start" link="/not-implemented" />
-          <SidebarItem label="Here" link="/not-implemented" />
-        </SidebarItem>
-      </Sidebar>
       <Router history={history}>
         <Suspense fallback={<div />}>
           <Switch>
