+++ src/components/content.js
@@ -21,14 +21,17 @@ const useStyles = makeStyles((theme) => ({
       easing: theme.transitions.easing.sharp,
       duration: theme.transitions.duration.leavingScreen,
     }),
-    marginLeft: -drawerWidth,
+    marginLeft: drawerWidth,
+    [theme.breakpoints.down('md')]: {
+      marginLeft: 0
+    }
   },
   contentShift: {
     transition: theme.transitions.create("margin", {
       easing: theme.transitions.easing.easeOut,
       duration: theme.transitions.duration.enteringScreen,
     }),
-    marginLeft: 0,
+    // marginLeft: 0,
   }
 }));
